Mostly working version for 1.21, except bugs

+++ src/main/java/dmr/DragonMounts/client/handlers/RidingCameraHandler.java
@@ -9,13 +9,13 @@ import net.minecraft.util.Mth;
 import net.minecraft.world.phys.AABB;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
-import net.neoforged.fml.common.Mod.EventBusSubscriber;
+import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.client.event.ViewportEvent.ComputeCameraAngles;
 
 @EventBusSubscriber( Dist.CLIENT )
 public class RidingCameraHandler
 {
-	private static double lastCameraIncrease;
+	private static float lastCameraIncrease;
 	
 	@SubscribeEvent
 	public static void flightCamera(ComputeCameraAngles setup){
@@ -24,14 +24,14 @@ public class RidingCameraHandler
 		
 		if (currentPlayer != null && currentPlayer.getRootVehicle() instanceof DMRDragonEntity) {
 			if (setup.getCamera().isDetached()) {
-				double gradualIncrease = Mth.lerp(0.25, lastCameraIncrease, info.getMaxZoom(10));
-				info.move(gradualIncrease * -1, (gradualIncrease / 2) * -1, 0);
+				float gradualIncrease = Mth.lerp(0.25f, lastCameraIncrease, info.getMaxZoom(10));
+				info.move(gradualIncrease * -1f, (gradualIncrease / 2f) * -1f, 0f);
 				lastCameraIncrease = gradualIncrease;
 			}
 		}
 		
 		if (lastCameraIncrease > 0) {
-			lastCameraIncrease = Mth.lerp(0.25, lastCameraIncrease, 0);
+			lastCameraIncrease = Mth.lerp(0.25f, lastCameraIncrease, 0);
 			info.move(0, lastCameraIncrease, 0);
 		}
 	}
